swap functions in MemPortUtils and MemTransformUtils properly for further refactoring
